ray_tracing: follow the memory2 -> memory package rename
