+++ dimos/mapping/ray_tracing/test_transformer.py
@@ -22,7 +22,7 @@ pytest.importorskip("dimos_voxel_ray_tracing")
 
 from dimos.mapping.ray_tracing.transformer import RayTraceMap
 from dimos.mapping.ray_tracing.voxel_map import VoxelRayMapper
-from dimos.memory2.type.observation import Observation
+from dimos.memory.type.observation import Observation
 from dimos.msgs.sensor_msgs.PointCloud2 import PointCloud2
